(refactor): remove unused assets and streamline form handling in Patient and VitalSigns components
