+++ HEALTH_TECH_FRONTEND/src/components/PatientForm/PatientForm.tsx
@@ -1,17 +1,9 @@
-import { useEffect } from 'react'
 import { useFormRegisterPacient } from '../../hooks/useFormRegisterPacient/useFormRegisterPacient'
 import type { PatientFormProps } from './types'
 
-export default function PatientForm({ onSuccess, onError }: PatientFormProps) {
-  const { form, loading, handleChange, handleKeyDown, handleSubmit, formError } = useFormRegisterPacient(onSuccess)
+export default function PatientForm({ onSuccess }: PatientFormProps) {
+  const { form, loading, handleChange, handleKeyDown, handleSubmit } = useFormRegisterPacient(onSuccess)
 
-  useEffect(() => {
-    if (formError && onError) {
-      const msg = Array.isArray(formError) ? formError.join(' · ') : formError
-      onError(msg)
-    }
-  }, [formError, onError])
-  
   return (
     <div className="bg-white-second-back rounded-lg p-6 w-full max-w-sm shadow-[0px_4px_16px_0px_rgba(0,0,0,0.2)]">
       <form className="flex flex-col gap-4" onSubmit={(e) => handleSubmit(e)}>
