+++ HEALTH_TECH_FRONTEND/src/components/PatientForm/types.ts
@@ -1,4 +1,4 @@
+
 export interface PatientFormProps {
-  onSuccess?: (e: React.FormEvent<HTMLFormElement>) => void
-  onError?: (message: string) => void
+  onSuccess?: (e: React.FormEvent<HTMLFormElement>) => void | Promise<void>
 }
