+++ HEALTH_TECH_FRONTEND/src/components/VitalSignsForm/types.ts
@@ -1,0 +1,3 @@
+export type VitalSignsFormProps = {
+  patientId: string
+}
